Fix parentCoordinator weak again
If the parentCoordinator would be deinitialized as well if its child was cancelled it could happen, that parentCoordinator is being deinitialized before its child is deinitialized. This would lead to a crash when doing parentCoordinator.removeChild(self).
As the child has no clue if its parent will survive its own cancellation it's saver to make the parentCoordinator weak instead of unowned. This makes this possible and therefore the code safer:
parentCoordinator?.removeChild(self)