Remove navigationController to be unowned.
Being unowned caused issues several time as some Coordinators should start with a fresh (presented) navigationController. This caused the issue, that something outside of the coordinator had to store the coordinator.

The best way to achieve this scenario would probably to have a PresetationCoordinator which could wait for a network call and then present the next coordinator which then starts with a fresh navigationController.
This also would help to use a coordinator presented and not presented multiple times without having to use different startup logic.